backend/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0
gunicorn==21.2.0
langchain==0.0.335
python-multipart==0.0.6
pytest==7.4.0
flask-cors==4.0.0
